CURB-6165 Refactored FavoritesButton component to ts
