+++ libraries/engage/components/v2/index.ts
@@ -1,5 +1,5 @@
 export { default as ButtonBase, type ButtonBaseProps } from './ButtonBase';
 export { default as Button, type ButtonProps } from './Button';
 export { default as ButtonGroup, type ButtonGroupProps } from './ButtonGroup';
-export { default as IconButton, type IconButtonProps } from './IconButton';
+export { default as IconButton, type IconButtonProps, type IconButtonSize } from './IconButton';
 export { default as CircularProgress, type CircularProgressProps } from './CircularProgress';
