+++ libraries/ui-shared/FavoritesButton/index.ts
@@ -1,0 +1,2 @@
+export { default } from './FavoritesButton';
+export type { FavoritesButtonOwnProps, FavoritesButtonProps } from './FavoritesButton';
